[inductor] Generate indirect_indexing checks even if optimized out
Previously `gen_assert_indirect_indexing` was only called on the index
expressions passed to `ops.load` and `ops.store` which means if the
variable is optimized out during lowering, we never generate the
assert. This instead makes `ops.indirect_indexing` eagerly generate
the assert statement, whether or not it will be used.

ghstack-source-id: ff6554222a4acca235b5f310aebad8e0017edbcf
Pull Request resolved: https://github.com/pytorch/pytorch/pull/100895